% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFileBrowser.R
\name{fileGetter}
\alias{fileGetter}
\title{Create a function that returns fileinfo according to the NOT given restrictions}
\usage{
fileGetter(dir2scan = getwd())
}
\arguments{
\item{dir2scan}{A logical value specifying whether hidden files should be
returned or not}
}
\value{
A function taking a single path relative to the specified root, and
returns a list of files to be passed on to shiny
}
\description{
Create a function that returns fileinfo according to the NOT given restrictions
}
