% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drives.R
\name{get_drives}
\alias{get_drives}
\title{Get the drives on the current machine}
\usage{
get_drives()
}
\value{
On Windows: the names of available drives. On non-Windows machines: the root
path \code{"/"} is returned.
}
\description{
On Windows machines, the standard drive is \verb{C:/}, but there may be other drives.
You can use this function to allow \code{\link[=file_browser]{file_browser()}} to browse in other drives.
}
\examples{
if (interactive()) {

### Simple app that lets you browse each drive

library(shiny)

ui <- fluidPage(
  selectInput("drive", "Select drive", get_drives()),
  file_browser_ui("files")
)

server <- function(input, output, session) {
  file_browser_server("files", path = reactive(input$drive))
}

shinyApp(ui, server)

##########################

### App that defaults to the home directory on C:/ drive

library(shiny)

ui <- fluidPage(
  selectInput("drive", "Select drive", get_drives()),
  file_browser_ui("files")
)

server <- function(input, output, session) {
  path <- reactive({
    if (input$drive == "C:/") {
      "~"
    } else {
      input$drive
    }
  })
  file_browser_server("files", path = path, root = reactive(input$drive))
}

shinyApp(ui, server)

}
}
