% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-browser.R
\name{paths_browser}
\alias{paths_browser}
\alias{path_browser_ui}
\alias{path_browser_server}
\title{Path browser}
\usage{
path_browser_ui(id, height = NULL)

path_browser_server(id, paths, show_extension = TRUE, parent_text = "..")
}
\arguments{
\item{id}{Unique ID for the module}

\item{height}{The height of the path browser. Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number which will be the number of pixels.}

\item{paths}{List of paths that the user can browse and select from. Use \code{/} as a
path separator. Any leading slashes are automatically removed.}

\item{show_extension}{(boolean) If \code{TRUE}, show file extensions in the file names.}

\item{parent_text}{The text to use to indicate the parent directory.}
}
\value{
List with reactive elements:
\itemize{
\item selected: The full normalized path of the selected file (\code{NULL} before a file is selected)
\item path: The full normalized path that is currently displayed in the file browser
}
}
\description{
Display a simple path browser that allows the user to browse paths and
select a path. This can also be used as a list selector.
}
