% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFileBrowser.R
\name{shinyFileBrowser}
\alias{shinyFileBrowser}
\title{Create a function that returns a tiny file browser}
\usage{
shinyFileBrowser(
  input,
  id,
  session = getSession(),
  message = "",
  width = NULL,
  height = NULL,
  elementId = paste0("shinyFileBrowser", as.character(as.integer(Sys.time()))),
  rootDirServer = NULL,
  rootDirHtml = "/",
  updateFreq = 0,
  ...
)
}
\arguments{
\item{input}{A}

\item{id}{A}

\item{session}{The session object of the shinyServer call (usually
`session`).}

\item{message}{A}

\item{width}{A}

\item{height}{A}

\item{elementId}{A}

\item{rootDirServer}{The default relative path specified given the `defaultRoot`.}

\item{rootDirHtml}{The default relative path specified given the `defaultRoot`.}

\item{updateFreq}{The default relative path specified given the `defaultRoot`.}

\item{...}{Arguments to be passed on to [fileGetter()]}
}
\value{
A reactive observer that takes care of the server side logic of the
filesystem connection.
}
\description{
Create a function that returns a tiny file browser
}
\concept{shinyFileBrowser}
