% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFileBrowser.R
\name{shinyFileBrowserExample}
\alias{shinyFileBrowserExample}
\title{Shiny bindings for shinyFileBrowser}
\usage{
shinyFileBrowserExample(rootDirServer = "~/", rootDirHtml = "/")
}
\arguments{
\item{rootDirServer}{output folder to read from}

\item{rootDirHtml}{output html of folder to read from}
}
\description{
Output and render functions for using shinyFileBrowser within Shiny
applications and interactive Rmd documents.
}
\seealso{
Other shinyFiles: 
\code{\link{shinyFileBrowser}()}
}
\concept{shinyFiles}
