% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-browser.R
\name{file_browser}
\alias{file_browser}
\alias{file_browser_ui}
\alias{file_browser_server}
\title{File browser}
\usage{
file_browser_ui(id, height = NULL, width = "100\%")

file_browser_server(
  id,
  path = getwd(),
  extensions = NULL,
  root = path,
  show_path = TRUE,
  show_extension = TRUE,
  show_size = TRUE,
  show_icons = TRUE,
  include_hidden = FALSE,
  include_empty = TRUE,
  parent_text = ".."
)
}
\arguments{
\item{id}{Unique ID for the module}

\item{height}{The height of the file browser. Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number which will be the number of pixels.}

\item{width}{The width of the file browser. Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number which will be the number of pixels.}

\item{path}{(reactive or static) The initial path the file browser should show.}

\item{extensions}{(reactive or static) List of file extensions that should be shown.
If \code{NULL}, all file types are shown.}

\item{root}{The path that should be considered root, which means that the user cannot
navigate to any parent of this root. By default, the \code{path} parameter acts as the root.
Use \code{NULL} to allow the user to navigate the entire filesystem.}

\item{show_path}{(boolean) If \code{TRUE}, show the current path above the browser.}

\item{show_extension}{(boolean) If \code{TRUE}, show file extensions in the file names.}

\item{show_size}{(boolean) If \code{TRUE}, show file sizes along the file names.}

\item{show_icons}{(boolean) If \code{TRUE}, show icons on the left column beside the file names.}

\item{include_hidden}{(boolean) If \code{TRUE}, show hidden files and folders.}

\item{include_empty}{(boolean) If \code{TRUE}, show empty files (files with size 0 bytes).}

\item{parent_text}{The text to use to indicate the parent directory.}
}
\value{
List with reactive elements:
\itemize{
\item selected: The full normalized path of the selected file (\code{NULL} before a file is selected)
\item path: The full normalized path that is currently displayed in the file browser
}
}
\description{
Display a simple file browser of the server-side file system.
}
